refacotr: migrate overview page to servers-component structure
